Shifted/Created components
CSS on App.vue was getting a little intense, and since it was the global css file, I just moved it to a separate file altogether.

Created the ItemCard component, which is basically the b-card from before on home, but as a separate component.

Created the ViewItem route, which views a single item.

Created the ItemSection component. For now, it is only used in ViewItem, but will be used in Home to view most of the items. Cards can be used in places, but the site looks odd with only cards.

Added the ViewItem route to routes.

Changed setCurrentItem (setCurrentEditItem before) in products store to an async function. I am going to try switching to async/await syntax where available.